requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
future==0.18.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
requests==2.22.0
urllib3==1.25.6
Werkzeug==0.16.0
